requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-cors-headers==3.5.0
django-rest-swagger==2.2.0
django-taggit==1.3.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
isort==5.6.1
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.4
Pillow==8.0.1
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pytz==2020.1
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1
